chain_of_thoughts/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
chain==1.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distlib==0.3.9
distro==1.9.0
filelock==3.17.0
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2024.10.1
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.34
langchain-openai==0.3.4
langchain-text-splitters==0.3.6
langsmith==0.3.8
lazy-object-proxy==1.10.0
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy>=1.26.4,<2
openai==1.61.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
orjson==3.10.15
packaging==24.2
pathable==0.4.4
pathspec==0.12.1
pipenv==2024.4.1
platformdirs==4.3.6
pluggy==1.5.0
prance==23.6.21.0
propcache==0.2.1
pyarrow==19.0.0
pyasn1==0.6.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pylance==0.23.0
PySimpleGUI==5.0.8
pytest==8.3.4
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.10
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.1
Werkzeug==3.1.3
yarl==1.18.3
zstandard==0.23.0
ipython>=8.0.0
nest_asyncio>=1.5.6

